backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
httpx
numpy
pypdf
pytesseract
pillow
passlib[bcrypt]
pytest
